todo/requirements.txt
asgiref==3.8.1
django-crispy-forms==2.3
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
